pyproject.toml
[tool.pdm]

[project]
name = "SpotiStats"
version = "0.1.0"
description = ""
authors = [
    {name = "jaypyles", email = "jpylesbusiness@gmail.com"},
]
dependencies = ['requests', 'python-dotenv', 'click']
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
